% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bohemiaR-package.R
\docType{package}
\name{bohemiaR-package}
\alias{bohemiaR}
\alias{bohemiaR-package}
\title{bohemiaR: Citas de Luces de Bohemia}
\description{
Colección de citas de la obra teatral Luces de bohemia (1924), de Ramón-María del Valle-Inclán.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pauinsanchez.github.io/bohemiaR/}
}

}
\author{
\strong{Maintainer}: Pau Insa-Sánchez \email{pau.insa@gmail.com} (\href{https://orcid.org/0000-0003-0841-7612}{ORCID})

}
\keyword{internal}
