% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dime_Latino.R
\name{Dime_Latino}
\alias{Dime_Latino}
\title{Dime Latino}
\usage{
Dime_Latino()
}
\value{
Una cadena de caracteres con una cita de don Latino, con indicación de la escena de la obra en la que aparece
}
\description{
Pide a don Latino de Híspalis que te deleite con una de sus frases de la obra
}
\examples{
Dime_Latino()
}
