% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dime_Max.R
\name{Dime_Max}
\alias{Dime_Max}
\title{Dime Max}
\usage{
Dime_Max()
}
\value{
Una cadena de caracteres con una cita de Max Estrella, con indicación de la escena de la obra en la que aparece
}
\description{
Pide a Max Estrella que te deleite con una de sus frases de la obra
}
\examples{
Dime_Max()
}
